Add Containers, Fix Semantics
Completely reworked the CSS to use containers to break up each element of the page, then sub-containers to actually hold content. Used containers to design elements of the page and flex with different viewscreen sizes, including mobile.

Added a navbar for mobile with a hamburger.

Reworked semantics to ensure that elements followed modern HTML standards, as some of the original work I did on the site a few years ago wasn't working well with the new design.

[This is a squash of other development commits]